pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hospital-er-triage-env"
version = "1.0.0"
description = "Hospital ER Triage Simulator — an OpenEnv environment"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"

dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["hospital_er_triage_env*", "server*"]

[project.scripts]
server = "server.app:main"
